+++ cmem_plugin_reason/plugin_validate.py
@@ -1,7 +1,6 @@
 """Ontology consistency validation workflow plugin module"""
 
 import shlex
-from collections.abc import Sequence
 from datetime import UTC, datetime
 from pathlib import Path
 from subprocess import run
@@ -22,7 +21,7 @@ from cmem_plugin_base.dataintegration.entity import (
 from cmem_plugin_base.dataintegration.plugins import WorkflowPlugin
 from cmem_plugin_base.dataintegration.types import BoolParameterType, StringParameterType
 from cmem_plugin_base.dataintegration.utils import setup_cmempy_user_access
-from pathvalidate import validate_filename
+from pathvalidate import is_valid_filename
 
 from cmem_plugin_reason.utils import (
     MAX_RAM_PERCENTAGE_DEFAULT,
@@ -33,6 +32,7 @@ from cmem_plugin_reason.utils import (
     ROBOT,
     create_xml_catalog_file,
     get_graphs_tree,
+    remove_temp,
     send_result,
 )
 
@@ -50,7 +50,8 @@ from cmem_plugin_reason.utils import (
             param_type=BoolParameterType(),
             name="write_md",
             label="Write Markdown explanation file",
-            description="Write Markdownn file with explanation to project.",
+            description="Write Markdown file with explanation to project. ⚠️ Existing files will "
+            "be overwritten.",
             default_value=False,
         ),
         PluginParameter(
@@ -65,7 +66,7 @@ from cmem_plugin_reason.utils import (
             name="output_graph_iri",
             label="Output graph IRI",
             description="The IRI of the output graph for the inconsistency validation. ⚠️ Existing "
-            "graph will be overwritten.",
+            "graphs will be overwritten.",
         ),
         PluginParameter(
             param_type=StringParameterType(),
@@ -85,7 +86,7 @@ from cmem_plugin_reason.utils import (
     ],
 )
 class ValidatePlugin(WorkflowPlugin):
-    """Example Workflow Plugin: Random Values"""
+    """Validate plugin"""
 
     def __init__(  # noqa: PLR0913
         self,
@@ -100,28 +101,26 @@ class ValidatePlugin(WorkflowPlugin):
     ) -> None:
         errors = ""
         if not validators.url(ontology_graph_iri):
-            errors += "Invalid IRI for parameter Ontology graph IRI. "
-        if reasoner not in REASONERS:
-            errors += "Invalid value for parameter Reasoner. "
+            errors += 'Invalid IRI for parameter "Ontology graph IRI." '
         if produce_graph and not validators.url(output_graph_iri):
-            errors += "Invalid IRI for parameter Output graph IRI. "
-        if write_md:
-            try:
-                validate_filename(md_filename)
-            except:  # noqa: E722
-                errors += "Invalid filename for parameter Output filename. "
-        if max_ram_percentage not in range(1, 100):
-            errors += "Invalid value for parameter Maximum RAM Percentage. "
+            errors += 'Invalid IRI for parameter "Output graph IRI". '
+        if produce_graph and output_graph_iri == ontology_graph_iri:
+            errors += "Output graph IRI cannot be the same as the Ontology graph IRI. "
+        if reasoner not in REASONERS:
+            errors += 'Invalid value for parameter "Reasoner". '
+        if write_md and not is_valid_filename(md_filename):
+            errors += 'Invalid filename for parameter "Output filename". '
+        if max_ram_percentage not in range(1, 101):
+            errors += 'Invalid value for parameter "Maximum RAM Percentage". '
         if errors:
             raise ValueError(errors[:-1])
-
         self.ontology_graph_iri = ontology_graph_iri
         self.reasoner = reasoner
         self.produce_graph = produce_graph
         self.output_graph_iri = output_graph_iri
         self.write_md = write_md
         self.stop_at_inconsistencies = stop_at_inconsistencies
-        self.md_filename = md_filename if md_filename and write_md else "mdfile.md"
+        self.md_filename = md_filename if write_md else "mdfile.md"
         self.max_ram_percentage = max_ram_percentage
         self.temp = f"reason_{uuid4().hex}"
 
@@ -176,35 +175,20 @@ class ValidatePlugin(WorkflowPlugin):
             replace=True,
         )
 
-    def clean_up(self, graphs: dict) -> None:
-        """Remove temporary files"""
-        files = ["catalog-v001.xml", "output.ttl", self.md_filename]
-        files += list(graphs.values())
-        for file in files:
-            try:
-                (Path(self.temp) / file).unlink()
-            except (OSError, FileNotFoundError) as err:
-                self.log.warning(f"Cannot remove file {file} ({err})")
-        try:
-            Path(self.temp).rmdir()
-        except (OSError, FileNotFoundError) as err:
-            self.log.warning(f"Cannot remove directory {self.temp} ({err})")
-
-    def execute(
-        self,
-        inputs: Sequence[Entities],  # noqa: ARG002
-        context: ExecutionContext,
-    ) -> Entities | None:
+    def execute(self, inputs: tuple, context: ExecutionContext) -> Entities | None:  # noqa: ARG002
         """Run the workflow operator."""
         setup_cmempy_user_access(context.user)
         graphs = get_graphs_tree((self.ontology_graph_iri,))
         self.get_graphs(graphs, context)
         create_xml_catalog_file(self.temp, graphs)
         self.validate(graphs)
+        files = ["catalog-v001.xml", self.md_filename, *graphs.values()]
+        if self.produce_graph:
+            files.append("output.ttl")
 
         text = (Path(self.temp) / self.md_filename).read_text()
         if text == "No explanations found.":
-            self.clean_up(graphs)
+            remove_temp(self, files)
             return None
 
         if self.produce_graph:
@@ -215,7 +199,7 @@ class ValidatePlugin(WorkflowPlugin):
             setup_cmempy_user_access(context.user)
             self.make_resource(context)
 
-        self.clean_up(graphs)
+        remove_temp(self, files)
 
         if self.stop_at_inconsistencies:
             raise RuntimeError("Inconsistencies found in Ontology.")